# openapi.yaml
openapi: 3.0.3
info:
  title: 人狼ゲームAPI（levi-api）
  description: |
    人狼ゲーム用のREST API仕様です。
    このAPIはユーザー管理、ゲーム管理、ゲーム進行、プレイヤーアクションなどの機能を提供します。
  version: 1.0.0
servers:
  - url: "https://api.gunjobiyori.com/v1"
    description: 本番環境
  - url: "https://staging-api.gunjobiyori.com/v1"
    description: ステージング環境
  - url: "http://localhost:8080/v1"
    description: 開発環境
paths:
...
  /users:
    post:
      tags:
        - users
      summary: 新規ユーザー登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: ユーザー登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /games:
    get:
      tags:
        - games
      summary: 利用可能なゲーム一覧を取得
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ゲーム一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      tags:
        - games
      summary: 新規ゲームを作成
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameCreation"
      responses:
        "201":
          description: ゲーム作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRegistration:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        stats:
          $ref: "#/components/schemas/UserStats"
    UserStats:
      type: object
      properties:
        gamesPlayed:
          type: integer
        gamesWon:
          type: integer
        winRatio:
          type: number
          format: float
        villagerWins:
          type: integer
        werewolfWins:
          type: integer
...
    GameCreation:
      type: object
      required:
        - name
        - maxPlayers
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 30
        maxPlayers:
          type: integer
          minimum: 3
          maximum: 20
        password:
          type: string
          description: オプションのゲームパスワード
        settings:
          $ref: "#/components/schemas/GameSettings"
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner:
          $ref: "#/components/schemas/User"
        hasPassword:
          type: boolean
        maxPlayers:
          type: integer
        currentPlayers:
          type: integer
        status:
          type: string
          enum: [WAITING, IN_PROGRESS, FINISHED]
        players:
          type: array
          items:
            $ref: "#/components/schemas/GamePlayer"
        createdAt:
          type: string
          format: date-time
        settings:
          $ref: "#/components/schemas/GameSettings"
        currentPhase:
          type: string
          enum: [DAY_DISCUSSION, DAY_VOTE, NIGHT, GAME_OVER]
          description: ゲームがIN_PROGRESS状態の時のみ有効
        currentDay:
          type: integer
          description: ゲーム日数（1日目から開始）
        phaseEndTime:
          type: string
          format: date-time
          description: 現在のフェーズ終了時刻
        winner:
          type: string
          enum: [VILLAGERS, WEREWOLVES, NONE]
          description: ゲーム終了時に設定される勝者陣営
        gameEvents:
          type: array
          items:
            $ref: "#/components/schemas/GameEvent"
    GameSettings:
      type: object
      properties:
        dayTimeSeconds:
          type: integer
          default: 300
          description: 昼フェーズの制限時間（秒）
        nightTimeSeconds:
          type: integer
          default: 180
          description: 夜フェーズの制限時間（秒）
        voteTimeSeconds:
          type: integer
          default: 60
          description: 投票フェーズの制限時間（秒）
        roles:
          $ref: "#/components/schemas/RoleSettings"
    RoleSettings:
      type: object
      properties:
        werewolfCount:
          type: integer
          default: 2
        seerCount:
          type: integer
          default: 1
        bodyguardCount:
          type: integer
          default: 1
        mediumCount:
          type: integer
          default: 0
    GamePlayer:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [VILLAGER, WEREWOLF, SEER, BODYGUARD, MEDIUM]
        isAlive:
          type: boolean
          default: true
        deathDay:
          type: integer
          description: プレイヤーが死亡した日
        deathCause:
          type: string
          enum: [WEREWOLF_ATTACK, EXECUTION, NONE]
    GameEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        day:
          type: integer
        phase:
          type: string
          enum: [DAY_DISCUSSION, DAY_VOTE, NIGHT]
        type:
          type: string
          enum: [PHASE_CHANGE, GAME_END]
        description:
          type: string
        timestamp:
          type: string
          format: date-time

    # アクション関連
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object